LayoutTests/fast/forms/input-step-number-expected.txt

Tests stepDown()/stepUp() for type=number.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Tests for an invalid current value:
PASS input.stepDown() threw exception Error: INVALID_STATE_ERR: DOM Exception 11.
PASS input.stepUp() threw exception Error: INVALID_STATE_ERR: DOM Exception 11.
stepDown("foo") with value=0:
PASS input.value is "0"
stepUp("foo") with value=0:
PASS input.value is "0"
stepDown(null) with value=0:
PASS input.value is "0"
stepUp(null) with value=0:
PASS input.value is "0"
stepDown(undefined) with value=0:
PASS input.value is "0"
stepUp(undefined) with value=0:
PASS input.value is "0"
stepUp() with value=0:
PASS input.value is "1"
stepUp(2) with value=1:
PASS input.value is "3"
stepUp(-1) with value=3:
PASS input.value is "2"
stepDown(-1) with value=2:
PASS input.value is "1"
stepDown(2) with value=1:
PASS input.value is "-1"
stepDown(-1) with value=-1:
PASS input.value is "0"
stepUp(1, 2) with value=0:
PASS input.value is "1"
stepUp(1, 3) with value=1:
PASS input.value is "0"
stepUp() with value=0 step=foo:
PASS input.value is "1"
stepUp() with value=1 step=0:
PASS input.value is "2"
stepUp() with value=2 step=-1:
PASS input.value is "3"
stepDown() and stepUp() with value=0 step=any:
PASS input.stepDown() threw exception Error: INVALID_STATE_ERR: DOM Exception 11.
PASS input.stepUp() threw exception Error: INVALID_STATE_ERR: DOM Exception 11.
stepDown() with value=1 min=0 step=1:
PASS input.value is "0"
stepDown() with value=0 min=0 step=1:
PASS input.stepDown() threw exception Error: INVALID_STATE_ERR: DOM Exception 11.
stepDown(2) with value=1 min=0 step=1:
PASS input.stepDown(2) threw exception Error: INVALID_STATE_ERR: DOM Exception 11.
PASS input.value is "1"
stepDown(2) with value=1 step=DBL_MAX:
PASS input.stepDown(2) threw exception Error: INVALID_STATE_ERR: DOM Exception 11.
stepUp() with value=-1 max=0 step=1:
PASS input.value is "0"
stepUp() with value=0 max=0 step=1:
PASS input.stepUp() threw exception Error: INVALID_STATE_ERR: DOM Exception 11.
stepUp(2) with value=-1 max=0 step=1:
PASS input.stepUp(2) threw exception Error: INVALID_STATE_ERR: DOM Exception 11.
PASS input.value is "-1"
stepUp(2) with value=1 step=DBL_MAX:
PASS input.stepUp(2) threw exception Error: INVALID_STATE_ERR: DOM Exception 11.
stepUp() with value=0 min=-1 step=2:
PASS input.value is "3"
PASS input.value is "1"
stepUp(9) with value=9 min=0 step=10:
PASS input.value is "100"
stepDown() with value=19 min=0 step=10:
PASS input.value is "10"
stepUp() with value=89 min=0 max=99 step=10:
PASS input.stepUp() threw exception Error: INVALID_STATE_ERR: DOM Exception 11.
stepUp(999999) with value=1e+308 step=1:
PASS input.value is "1e+308"
PASS input.value is "1e+308"
stepUp(3) with value=0 step=0.33333333333333333:
PASS input.value is "1"
stepUp(10) with value=1 step=0.1:
PASS input.value is "2"
stepUp() x10 with value=2 step=0.1:
PASS input.value is "3"
stepUp(255) with value=0 min=0 max=1 step=0.003921568627450980:
PASS input.value is "1"
stepDown() x255 with value=1 min=0 max=1 step=0.003921568627450980:
PASS input.value is "0"
PASS successfullyParsed is true

TEST COMPLETE
